// AgentFront/src/app/model/new-unit-dto.ts
export class NewUnitDto {
    capacity:Number;
    canBeCanceled:boolean;
    daysBefore:Number;
    january: Number;
    february: Number;
    march: Number;
    april: Number;
    may: Number;
    june: Number;
    july: Number;
    august: Number;
    september: Number;
    october: Number;
    november: Number;
    december: Number;
    facilityId:Number;
    unitAS:Number[];
}